requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.1
django-celery-beat==2.5.0
django-guardian==2.4.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
django-filter==23.3
drf-yasg==1.21.7
python-json-logger==2.0.7
prometheus-client==0.19.0
cryptography==41.0.7